add base data value format
